move hooksmixin to separate file
