+++ src/llmcompressor/transformers/finetune/data/data_helpers.py
@@ -27,10 +27,11 @@ def create_single_batch_dataloader(
 ) -> torch.utils.data.DataLoader:
     """
     Create a dataloader whose batch size is equal to the size of the dataset
-    
+
     :param dataset: dataset used to generate dataloader
     :return: dataloader
     """
+
     def pad_sequences(batch):
         # extract input_ids and attention_mask from the batch
         input_ids = [torch.tensor(item["input_ids"]) for item in batch]
